// src/models/IoK8sApiApiserverinternalV1alpha1StorageVersionList/types.ts
import { IoK8sApiApiserverinternalV1alpha1StorageVersionModelOptions as IoK8sApiApiserverinternalV1alpha1StorageVersion } from '../IoK8sApiApiserverinternalV1alpha1StorageVersion';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiApiserverinternalV1alpha1StorageVersionListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiApiserverinternalV1alpha1StorageVersion>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}